python/samples/web_browsing_agent/pyproject.toml
[project]
name = "web-browsing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.3",
    "browser-use>=0.1.36",
    "python-dotenv>=1.0.1",
    "teams-ai==1.5.0"
]
